requirements.txt
numpy
tqdm
pillow
scikit-learn
torch
torchvision
tensorboardX
matplotlib
opencv-python
dlib
scikit-build
